feat(auth): authenticable entity seeder
