+++ packages/core/manifest/src/manifest/controllers/manifest.controller.ts
@@ -4,6 +4,7 @@ import { Request } from 'express'
 import { AuthService } from '../../auth/auth.service'
 import { ManifestService } from '../services/manifest.service'
 import { ApiTags } from '@nestjs/swagger'
+import { ADMIN_ENTITY_MANIFEST } from '../../constants'
 
 @ApiTags('Manifest')
 @Controller('manifest')
@@ -14,20 +15,21 @@ export class ManifestController {
   ) {}
 
   @Get()
-  async getPublicManifest(@Req() req: Request): Promise<AppManifest> {
+  async getManifest(@Req() req: Request): Promise<AppManifest> {
+    // TODO: Make this cleaner (and below)
     const currentUser: AuthenticableEntity =
-      await this.authService.getUserFromRequest(req, 'admins')
+      await this.authService.getUserFromRequest(req, ADMIN_ENTITY_MANIFEST.slug)
 
     return this.manifestService.getAppManifest({ publicVersion: !currentUser })
   }
 
   @Get('entities/:slug')
-  async getEntityPublicManifest(
+  async getEntityManifest(
     @Param('slug') slug: string,
     @Req() req: Request
   ): Promise<EntityManifest> {
     const currentUser: AuthenticableEntity =
-      await this.authService.getUserFromRequest(req, 'admins')
+      await this.authService.getUserFromRequest(req, ADMIN_ENTITY_MANIFEST.slug)
 
     return this.manifestService.getEntityManifest({
       slug,
